fix(android): reduce DoS surface, consolidate DPM lookup, extract shared FakeIntent
- short-circuit SamsungHardKeyReceiver.onReceive on unknown broadcast
  actions before any DevicePolicyManager binder call; this removes
  unnecessary IPC for every arbitrary broadcast sent to the exported
  receiver by other apps
- fetch DevicePolicyManager once and pass the result to both
  isDeviceOwnerApp and isProfileOwnerApp instead of retrieving the
  system service twice via separate private helpers
- extract the shared FakeIntent intent stub into a package-private
  class so SamsungHardKeyReceiverTest and SamsungHardwareButtonLaunchTest
  no longer duplicate the implementation; the extracted class is a
  superset of both (action + componentName + all extras)
- add ignoresUnknownActionBroadcastsEvenInManagedMode test to document
  the filtering invariant explicitly
