+++ android/app/src/test/java/app/secpal/SamsungHardKeyReceiverTest.java
@@ -10,9 +10,6 @@ import static org.junit.Assert.assertNull;
 
 import android.content.Intent;
 
-import java.util.HashMap;
-import java.util.Map;
-
 import org.junit.Test;
 
 public class SamsungHardKeyReceiverTest {
@@ -26,6 +23,20 @@ public class SamsungHardKeyReceiverTest {
                 intent,
                 "app.secpal",
                 false,
+                false
+            )
+        );
+    }
+
+    @Test
+    public void ignoresUnknownActionBroadcastsEvenInManagedMode() {
+        FakeIntent intent = new FakeIntent("com.example.unrelated.action");
+
+        assertNull(
+            SamsungHardKeyReceiver.resolveManagedHardwareAction(
+                intent,
+                "app.secpal",
+                true,
                 false
             )
         );
@@ -53,69 +64,5 @@ public class SamsungHardKeyReceiverTest {
                 true
             )
         );
-    }
-
-    private static final class FakeIntent extends Intent {
-        private final Map<String, Object> extras = new HashMap<>();
-        private String action;
-
-        private FakeIntent(String action) {
-            this.action = action;
-        }
-
-        @Override
-        public String getAction() {
-            return action;
-        }
-
-        @Override
-        public Intent setAction(String action) {
-            this.action = action;
-            return this;
-        }
-
-        @Override
-        public boolean hasExtra(String name) {
-            return extras.containsKey(name);
-        }
-
-        @Override
-        public String getStringExtra(String name) {
-            Object value = extras.get(name);
-
-            return value instanceof String ? (String) value : null;
-        }
-
-        @Override
-        public boolean getBooleanExtra(String name, boolean defaultValue) {
-            Object value = extras.get(name);
-
-            return value instanceof Boolean ? ((Boolean) value).booleanValue() : defaultValue;
-        }
-
-        @Override
-        public int getIntExtra(String name, int defaultValue) {
-            Object value = extras.get(name);
-
-            return value instanceof Integer ? ((Integer) value).intValue() : defaultValue;
-        }
-
-        @Override
-        public Intent putExtra(String name, String value) {
-            extras.put(name, value);
-            return this;
-        }
-
-        @Override
-        public Intent putExtra(String name, boolean value) {
-            extras.put(name, Boolean.valueOf(value));
-            return this;
-        }
-
-        @Override
-        public Intent putExtra(String name, int value) {
-            extras.put(name, Integer.valueOf(value));
-            return this;
-        }
     }
 }
