+++ android/app/src/test/java/app/secpal/SamsungHardwareButtonLaunchTest.java
@@ -13,9 +13,6 @@ import static org.junit.Assert.assertTrue;
 import android.content.ComponentName;
 import android.content.Intent;
 
-import java.util.HashMap;
-import java.util.Map;
-
 import org.junit.Before;
 import org.junit.Test;
 
@@ -248,84 +245,5 @@ public class SamsungHardwareButtonLaunchTest {
             SamsungHardwareButtonLaunch.HARDWARE_TRIGGER_ACTION_LONG_PRESS,
             SamsungHardwareButtonLaunch.resolveLaunchAction(upIntent, "app.secpal")
         );
-    }
-
-    private static final class FakeIntent extends Intent {
-        private final Map<String, Object> extras = new HashMap<>();
-        private String action;
-        private ComponentName componentName;
-
-        private FakeIntent() {
-        }
-
-        private FakeIntent(String action) {
-            this.action = action;
-        }
-
-        @Override
-        public String getAction() {
-            return action;
-        }
-
-        @Override
-        public Intent setAction(String action) {
-            this.action = action;
-            return this;
-        }
-
-        @Override
-        public ComponentName getComponent() {
-            return componentName;
-        }
-
-        @Override
-        public Intent setComponent(ComponentName componentName) {
-            this.componentName = componentName;
-            return this;
-        }
-
-        @Override
-        public boolean hasExtra(String name) {
-            return extras.containsKey(name);
-        }
-
-        @Override
-        public String getStringExtra(String name) {
-            Object value = extras.get(name);
-
-            return value instanceof String ? (String) value : null;
-        }
-
-        @Override
-        public boolean getBooleanExtra(String name, boolean defaultValue) {
-            Object value = extras.get(name);
-
-            return value instanceof Boolean ? ((Boolean) value).booleanValue() : defaultValue;
-        }
-
-        @Override
-        public int getIntExtra(String name, int defaultValue) {
-            Object value = extras.get(name);
-
-            return value instanceof Integer ? ((Integer) value).intValue() : defaultValue;
-        }
-
-        @Override
-        public Intent putExtra(String name, String value) {
-            extras.put(name, value);
-            return this;
-        }
-
-        @Override
-        public Intent putExtra(String name, boolean value) {
-            extras.put(name, Boolean.valueOf(value));
-            return this;
-        }
-
-        @Override
-        public Intent putExtra(String name, int value) {
-            extras.put(name, Integer.valueOf(value));
-            return this;
-        }
     }
 }
